Rename GlobalLoading file to snake-case.
