+++ app/assets/scripts/components/common/global-loading.js
